add --no_FS_codon option when calling Extract_sequences_from_genome.py
Instead of adding a X for every frameshift, the sequence remains unchanged (and so does its length).
This possibly changes hmm profiles and LRR domains detection. On the DW dataset it was tested on, 5 genes became LRRs and 5 genes were no longer LRRs compared to before this change.
We looked at a gene that was no longer classified as an LRR (0 motifs) with this current version, and found that it had previously been annotated with 9 LRR motifs by LRRprofiler using the previous version with X's in the input sequences.